feat: add interactive questions message type
Add a first-class multi-turn question card message type so agents can ask
structured choice questions and the mobile client renders clickable
single/multi-select UI instead of plain text.

Agent emits via new MCP tool orchestrator__ask_questions; answers flow
back over Socket.IO (message.answers). On submission the backend
persists content.answers, broadcasts message.updated to flip the card
read-only, then synthesizes a user text message so the conversation
continues through the normal streaming pipeline.

Also flip all mobile-proto int64 counters/pagination/size fields to
int32 so proto3 JSON encodes them as numbers (not strings), fixing a
Dart type-cast crash in the mobile client.

+++ internal/orchestrator/server/handler/usage.go
@@ -37,11 +37,11 @@ func UserUsageSummary(c *Context) http.HandlerFunc {
 
 		WriteProtoSuccess(w, http.StatusOK, &mobilev1.UserUsageSummaryResponse{
 			CurrentPeriod:        period,
-			TokensUsed:           tokensUsed,
-			PromptTokensUsed:     counters.PromptTokensUsed,
-			CompletionTokensUsed: counters.CompletionTokensUsed,
-			RequestCount:         int64(counters.RequestCount),
-			TokenLimit:           tokenLimit,
+			TokensUsed:           int32(tokensUsed),                    // #nosec G115 -- token count within int32 range for display
+			PromptTokensUsed:     int32(counters.PromptTokensUsed),     // #nosec G115 -- token count within int32 range for display
+			CompletionTokensUsed: int32(counters.CompletionTokensUsed), // #nosec G115 -- token count within int32 range for display
+			RequestCount:         int32(counters.RequestCount),         // #nosec G115 -- request count fits in int32
+			TokenLimit:           int32(tokenLimit),                    // #nosec G115 -- token limit within int32 range for display
 			PlanId:               counters.PlanID,
 		})
 	}
@@ -83,11 +83,11 @@ func WorkspaceUsage(c *Context) http.HandlerFunc {
 
 		WriteProtoSuccess(w, http.StatusOK, &mobilev1.WorkspaceUsageResponse{
 			Period:               period,
-			TokensUsed:           counters.TokensUsed,
-			PromptTokensUsed:     counters.PromptTokensUsed,
-			CompletionTokensUsed: counters.CompletionTokensUsed,
-			RequestCount:         int64(counters.RequestCount),
-			TokenLimit:           tokenLimit,
+			TokensUsed:           int32(counters.TokensUsed),           // #nosec G115 -- token count within int32 range for display
+			PromptTokensUsed:     int32(counters.PromptTokensUsed),     // #nosec G115 -- token count within int32 range for display
+			CompletionTokensUsed: int32(counters.CompletionTokensUsed), // #nosec G115 -- token count within int32 range for display
+			RequestCount:         int32(counters.RequestCount),         // #nosec G115 -- request count fits in int32
+			TokenLimit:           int32(tokenLimit),                    // #nosec G115 -- token limit within int32 range for display
 		})
 	}
 }
